backend/requirements.txt
flask>=3.0,<4
flask-cors>=5.0,<6
sqlalchemy>=2.0.36,<3
python-jose[cryptography]>=3.3.0,<4
bcrypt>=4.0,<5
pydantic[email]>=2.11.0,<3
pydantic-settings>=2.6,<3
certifi>=2024.0.0
psycopg[binary]>=3.2,<4
pytest>=8.0,<9
